wip: opt-in ambient
